pyproject.toml
[tool.poetry]
name = "telegramttsbot"
version = "0.1.0"
description = ""
authors = ["lriazy <leili.riazy@charite.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.12.0"
google-cloud-secret-manager = "^2.20.2"
vertexai = "^1.63.0"
google-cloud-aiplatform = "^1.63.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
